+++ src/ls/driver.ts
@@ -367,7 +367,7 @@ export default class MSSQL
           {
             label: "Linked Servers",
             database: "Linked Servers",
-            type: "connection.linkedServers",
+            type: "connection.linkedServersRoot",
             detail: "linkedservers",
             iconId: "link",
           },
@@ -446,7 +446,6 @@ export default class MSSQL
             iconId: "folder",
             childType: ContextValue.VIEW,
           },
-          // { label: 'Functions', type: ContextValue.RESOURCE_GROUP, iconId: 'folder', childType: ContextValue.FUNCTION },
         ];
       case ContextValue.FUNCTION:
         return <MConnectionExplorer.IChildItem[]>(<unknown>[
@@ -476,7 +475,7 @@ export default class MSSQL
         return this.queryResults(this.queries.fetchParameters(item));
       case "connection.storedProceduresRoot":
         return this.queryResults(this.queries.fetchStoredProcedures());
-      case "connection.linkedServers":
+      case "connection.linkedServersRoot":
         try {
           const result = await this.queryResults(
             this.queries.fetchLinkedServers()
@@ -486,6 +485,136 @@ export default class MSSQL
           this.close();
           return [];
         }
+      case "connection.linkedServers":
+        try {
+          const result = await this.queryResults(
+            this.queries.fetchLinkedServerCatalogs(item)
+          );
+          const transformedResult = result.map((out) => ({
+            label: out.CATALOG_NAME,
+            database: out.CATALOG_NAME,
+            type: ContextValue.DATABASE,
+            detail: "database",
+            linkedserver: item.linkedserver,
+          }));
+          return transformedResult;
+        } catch (error) {
+          console.log(
+            "Error while fetching linked servers:",
+            error,
+            "\n----Using a fallback method----"
+          );
+          try {
+            const result = await this.queryResults(
+              this.queries.fetchLinkedServerDatabaseAndSchemas(item)
+            );
+            const distinctValues = Array.from(
+              new Set(
+                result.map((item) => `${item.TABLE_CAT}:${item.TABLE_SCHEM}`)
+              )
+            ).map((item) => {
+              const [TABLE_CAT, TABLE_SCHEM] = item.split(":");
+              return { TABLE_CAT, TABLE_SCHEM };
+            });
+            const fallbackDatabases = distinctValues.map((out) => ({
+              label: out.TABLE_CAT,
+              database: out.TABLE_CAT,
+              schema: out.TABLE_SCHEM,
+              type: "connection.fallbackDatabase",
+              detail: "database",
+              linkedserver: item.linkedserver,
+              iconId: "database",
+            }));
+            const fallbackSchemas = distinctValues.map((out) => ({
+              label: out.TABLE_SCHEM,
+              database: out.TABLE_CAT,
+              schema: out.TABLE_SCHEM,
+              type: "connection.fallbackSchema",
+              detail: "schema",
+              linkedserver: item.linkedserver,
+              iconId: "group-by-ref-type",
+            }));
+            if (!result.every((row) => row.TABLE_CAT === null)) {
+              return fallbackDatabases;
+            } else {
+              return fallbackSchemas;
+            }
+          } catch (error) {
+            console.error("Error in fallback method:", error);
+            this.close();
+            return [];
+          }
+        }
+      case "connection.fallbackDatabase":
+        try {
+          const result = await this.queryResults(
+            this.queries.fetchLinkedServerSchemas(item)
+          );
+          const distinctSchem = [
+            ...new Set(result.map((row) => row.TABLE_SCHEM)),
+          ];
+          const distinctSchemObjects = distinctSchem.map((schem) => ({
+            TABLE_SCHEM: schem,
+          }));
+          const fallbackSchemas = distinctSchemObjects.map((out) => ({
+            label: out.TABLE_SCHEM,
+            database: item.database,
+            schema: out.TABLE_SCHEM,
+            type: "connection.fallbackSchema",
+            detail: "schema",
+            linkedserver: item.linkedserver,
+            iconId: "group-by-ref-type",
+          }));
+          return fallbackSchemas;
+        } catch (error) {
+          this.close();
+          return [];
+        }
+      case "connection.fallbackSchema":
+        return [
+          {
+            label: "Tables",
+            type: ContextValue.RESOURCE_GROUP,
+            iconId: "folder",
+            childType: "connection.fallbackTable",
+          },
+          {
+            label: "Views",
+            type: ContextValue.RESOURCE_GROUP,
+            iconId: "folder",
+            childType: "connection.fallbackView",
+          },
+        ];
+      case "connection.fallbackTable":
+      case "connection.fallbackView":
+        let fallbackColumnResult = await this.queryResults(
+          this.queries.fetchLinkedServerColumns(item)
+        );
+        fallbackColumnResult = fallbackColumnResult.filter(
+          (column) =>
+            column.TYPE_NAME !== "INTERVAL YEAR TO MONTH" &&
+            column.TYPE_NAME !== "INTERVAL DAY TO SECOND"
+        );
+        const transformedColumns = fallbackColumnResult.map((column) => {
+          const detail = column.CHAR_OCTET_LENGTH
+            ? `${column.TYPE_NAME}(${column.CHAR_OCTET_LENGTH})`
+            : column.TYPE_NAME;
+
+          return {
+            label: column.COLUMN_NAME,
+            type: "connection.fallbackColumn",
+            table: column.TABLE_NAME,
+            dataType: column.TYPE_NAME,
+            detail: detail,
+            database: column.TABLE_CAT,
+            schema: column.TABLE_SCHEM,
+            isNullable: column.IS_NULLABLE,
+            iconId: "symbol-field",
+            childType: ContextValue.NO_CHILD,
+          };
+        });
+
+        return transformedColumns;
     }
     return [];
   }
@@ -508,10 +637,15 @@ export default class MSSQL
     table: NSDatabase.ITable,
     opt: IQueryOptions & { limit: number; page?: number }
   ) {
-    const sqlServerVersionQuery = await this.query(
-      "SELECT SERVERPROPERTY('ProductVersion') AS Version",
-      {}
-    );
+    const sqlServerVersionQuery = table.linkedserver
+      ? await this.query(
+          `SELECT * FROM OPENQUERY(${table.linkedserver}, 'SELECT SERVERPROPERTY(''productversion'') AS Version')`,
+          {}
+        )
+      : await this.query(
+          "SELECT SERVERPROPERTY('ProductVersion') AS Version",
+          {}
+        );
     const sqlServerVersion = sqlServerVersionQuery[0].results[0].Version;
     // SQL Server 2012 added support for OFFSET
     if (this.compareVersions(sqlServerVersion, "11.0.1103.9")) {
@@ -592,6 +726,38 @@ export default class MSSQL
           ...col,
           childType: ContextValue.NO_CHILD,
         }));
+      case "connection.fallbackTable":
+        const resultsFallbackTables = await this.queryResults(
+          this.queries.fetchLinkedServerTables(parent)
+        );
+        const transformedResultsFallbackTables = resultsFallbackTables.map(
+          (out) => ({
+            label: out.TABLE_NAME,
+            database: out.TABLE_CAT,
+            schema: out.TABLE_SCHEM,
+            type: "connection.fallbackTable",
+            isView: 0,
+            linkedserver: parent.linkedserver,
+            iconName: "table",
+          })
+        );
+        return transformedResultsFallbackTables;
+      case "connection.fallbackView":
+        const resultsFallbackViews = await this.queryResults(
+          this.queries.fetchLinkedServerViews(parent)
+        );
+        const transformedResultsFallbackViews = resultsFallbackViews.map(
+          (out) => ({
+            label: out.TABLE_NAME,
+            database: out.TABLE_CAT,
+            schema: out.TABLE_SCHEM,
+            type: "connection.fallbackView",
+            isView: 1,
+            linkedserver: parent.linkedserver,
+            iconName: "view",
+          })
+        );
+        return transformedResultsFallbackViews;
     }
     return [];
   }